fix(#632): fixed random deadlock while training
While creating new blocks in the training-process
a deadlock could appear. This should now be fixed
by restructureing the resizing functions.
