+++ src/binaries/hanami/src/core/blocks/mod.rs
@@ -19,3 +19,4 @@ pub mod new_core_block;
 pub mod axons;
 pub mod block_trait;
 pub mod block_io;
+pub mod target_search;
